First auth step successfully implemented AppKey
Where could I start? trying to cipher the appKey, avoiding that any
variable (apart from the api response body obviously)
will store the appKey I figer out tha the Gson converter
takes a deafault setter no matter that I declared a new one,
was necesary implement a new dependency

kotlinx-serialization-json:1.6.3

also do some changes to get the context for create the
secretFiles or in the Login view to show a Toast
if something went wrong with the  login process

Also all the logic for the api call was implemented
the base Url has been saved in the strings in order
to get a eassy control of this.

Sorry for the long commit, Ill try this doesnt happend again